checkout: extract inline SVGs into components/icons
