+++ clients/packages/checkout/src/components/MeteredPricesDisplay.tsx
@@ -9,26 +9,7 @@ import { useMemo } from 'react'
 import { ProductCheckoutPublic } from '../guards'
 import { getMeteredPrices } from '../utils/product'
 import ProductPriceLabel from './ProductPriceLabel'
-
-const GaugeIcon = ({ className }: { className?: string }) => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      className={className}
-    >
-      <path d="m12 14 4-4" />
-      <path d="M3.34 19a10 10 0 1 1 17.32 0" />
-    </svg>
-  )
-}
+import GaugeIcon from './icons/GaugeIcon'
 
 interface MeteredPricesDisplayProps {
   checkout: ProductCheckoutPublic
