+++ clients/packages/checkout/src/components/icons/PlusIcon.tsx
@@ -1,0 +1,20 @@
+const PlusIcon = ({ className }: { className?: string }) => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 14 14"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      aria-hidden="true"
+      focusable="false"
+      className={className}
+    >
+      <path d="M7 3v8M3 7h8" />
+    </svg>
+  )
+}
+
+export default PlusIcon
